[DOP-4294]: Add tests
